Rename sample.scalagwt.HelloWorld to sample.jribble.Hello.
This way we can make use of fake jribble parser that is already committed
to GWT repository in order to make gwtc target working.

This fake parser produces constant (hard-coded) AST disregarding input it
was passed. We need to rename our project in order to be in-line with
hard-coded values.

Another benefit is that original GWT sample is called hello and not hello
world so it's better to reuse the name.

This commit makes gwtc 'working' in a sense that it doesn't fail but it
still doesn't process the input jribble files it's given.

Change-Id: I772638b71569f17609dba2846d1ccf7bd885794f
Signed-off-by: Grzegorz Kossakowski <grek@google.com>