Update classpath settings in samples.
Many updates to classpath settings for compiling
samples to accommodate to changes in
scalagwt-scala project.

Specific changes include:

  * use factorymanifests plug-in
  * use scala-library-gwt.jar in approporiate places
  * separate compilation of generators from other
    source code

The most questionable bullet is the last one. The reason
why we need separate compilation of generators is because
they need to be compiled against scala-library.jar and
not scala-library-gwt.jar. That comes from the fact that
GWT is going to load compiled generators and try to run
them using it's standard classpath that includes
scala-library.jar.

The best solution would be to have separate, compilation
classpath implemented in GWT. Until then we need to
be very careful about classpath used for compilation of
different parts of source code.

Change-Id: I22fcf91eaed89a5eef17920724f11d6fd97b9f62
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>