Cleanups to build.xml file.
Removed unused targets.

Added greetings target that greets brave souls
that decided to try out what we have. Moreover,
we warn people that compilation process is very
slow at the moment so they are less surprised
by that. That's a good chance to let them know
that we want to fix it in near future, too.

Change-Id: Ie417dbb2547d2664a454425fc979db40c6e6aab1
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>